+++ index.js
@@ -1,5 +1,5 @@
-const backup = require('./lib/wal-e_backup');
-const restore = require('./lib/wal-e_restore');
+const backup = require('./lib/backup');
+const restore = require('./lib/restore');
 const slack = require('./lib/slack');
 const gcs = require('./lib/gcs');
 const logger = require('./lib/logging');
