+++ lib/gcs.js
@@ -3,7 +3,7 @@ const moment = require('moment');
 const promisify = require('util').promisify;
 const exec = promisify(require('child_process').exec);
 const logger = require('./logging');
-const {getBackups} = require('./wal-e_backup');
+const {getBackups} = require('./backup');
 
 const wale = cfg('wale');
 const gsutil = '/usr/local/bin/gsutil';
