+++ lib/restore.js
@@ -5,7 +5,7 @@ const promisify = require('util').promisify;
 const {spawn, exec: origExec} = require('child_process');
 
 const logger = require('./logging');
-const {options, getBackups, getGsPrefix, customSpawn} = require('./wal-e_backup');
+const {options, getBackups, getGsPrefix, customSpawn} = require('./backup');
 
 const exec = promisify(origExec);
 const fs = {
